Reorganize the mule package classes
