+++ mule/src/main/java/mule/MuleToBalConverter.java
@@ -21,7 +21,8 @@ import common.BallerinaModel.TypeDesc.RecordTypeDesc;
 import common.BallerinaModel.TypeDesc.RecordTypeDesc.RecordField;
 import common.CodeGenerator;
 import io.ballerina.compiler.syntax.tree.SyntaxTree;
-import mule.MuleXMLNavigator.MuleElement;
+import mule.reader.MuleXMLNavigator;
+import mule.reader.MuleXMLNavigator.MuleElement;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.xml.sax.SAXException;
@@ -63,23 +64,23 @@ import static mule.ConversionUtils.getBallerinaAbsolutePath;
 import static mule.ConversionUtils.getBallerinaResourcePath;
 import static mule.ConversionUtils.getAttrValInt;
 import static mule.ConversionUtils.insertLeadingSlash;
-import static mule.MuleConfigConverter.convertTopLevelMuleBlocks;
-import static mule.MuleConfigConverter.getCatchExceptionBody;
-import static mule.MuleConfigConverter.getChoiceExceptionBody;
-import static mule.MuleConfigReader.readMuleConfigFromRoot;
-import static mule.MuleModel.CatchExceptionStrategy;
-import static mule.MuleModel.ChoiceExceptionStrategy;
-import static mule.MuleModel.DbMSQLConfig;
-import static mule.MuleModel.DbOracleConfig;
-import static mule.MuleModel.DbTemplateQuery;
-import static mule.MuleModel.Flow;
-import static mule.MuleModel.HTTPListenerConfig;
-import static mule.MuleModel.HttpListener;
-import static mule.MuleModel.Kind;
-import static mule.MuleModel.MuleRecord;
-import static mule.MuleModel.SubFlow;
-import static mule.MuleModel.VMInboundEndpoint;
-import static mule.MuleModel.UnsupportedBlock;
+import static mule.converter.MuleConfigConverter.convertTopLevelMuleBlocks;
+import static mule.converter.MuleConfigConverter.getCatchExceptionBody;
+import static mule.converter.MuleConfigConverter.getChoiceExceptionBody;
+import static mule.reader.MuleConfigReader.readMuleConfigFromRoot;
+import static mule.model.MuleModel.CatchExceptionStrategy;
+import static mule.model.MuleModel.ChoiceExceptionStrategy;
+import static mule.model.MuleModel.DbMSQLConfig;
+import static mule.model.MuleModel.DbOracleConfig;
+import static mule.model.MuleModel.DbTemplateQuery;
+import static mule.model.MuleModel.Flow;
+import static mule.model.MuleModel.HTTPListenerConfig;
+import static mule.model.MuleModel.HttpListener;
+import static mule.model.MuleModel.Kind;
+import static mule.model.MuleModel.MuleRecord;
+import static mule.model.MuleModel.SubFlow;
+import static mule.model.MuleModel.VMInboundEndpoint;
+import static mule.model.MuleModel.UnsupportedBlock;
 
 public class MuleToBalConverter {
