+++ mule/src/main/java/mule/dataweave/converter/DWReader.java
@@ -23,9 +23,9 @@ import common.BallerinaModel.Statement.BallerinaStatement;
 import mule.Constants;
 import mule.Context;
 import mule.ConversionUtils;
-import mule.MuleModel;
 import mule.dataweave.parser.DataWeaveLexer;
 import mule.dataweave.parser.DataWeaveParser;
+import mule.model.MuleModel;
 import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.tree.ParseTree;
