+++ mule/src/main/java/mule/model/MuleModel.java
@@ -15,7 +15,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package mule;
+package mule.model;
 
 import java.util.List;
 import java.util.Map;
